feat: move vrender-animate/interface to vrender-core
