+++ packages/vrender-animate/src/custom/clip-graphic.ts
@@ -1,7 +1,6 @@
 import type { IArcGraphicAttribute, IGraphic, IGroup, IRectGraphicAttribute } from '@visactor/vrender-core';
-import { application, AttributeUpdateType } from '@visactor/vrender-core';
+import { application, AttributeUpdateType, type EasingType } from '@visactor/vrender-core';
 import { ACustomAnimate } from './custom-animate';
-import type { EasingType } from '../intreface/easing';
 
 export class ClipGraphicAnimate extends ACustomAnimate<any> {
   private _group?: IGroup;
