+++ packages/vrender-animate/src/custom/richtext/slide-richtext.ts
@@ -1,7 +1,5 @@
-import type { IAnimate, IStep } from '../../intreface/animate';
-import type { EasingType } from '../../intreface/easing';
 import { ACustomAnimate } from '../custom-animate';
-import type { IRichTextCharacter, IRichTextParagraphCharacter } from '@visactor/vrender-core';
+import type { IRichTextCharacter, IAnimate, IStep, EasingType } from '@visactor/vrender-core';
 import { RichText } from '@visactor/vrender-core';
 
 /**
