+++ packages/vrender-components/src/axis/animate/group-transition.ts
@@ -1,5 +1,6 @@
+import { AnimateMode } from '@visactor/vrender-core';
 import type { EasingType, IGraphic, IGroup } from '@visactor/vrender-core';
-import { ACustomAnimate, AnimateMode } from '@visactor/vrender-animate';
+import { ACustomAnimate } from '@visactor/vrender-animate';
 import type { Dict } from '@visactor/vutils';
 import { cloneDeep, interpolateString, isEqual, isValidNumber } from '@visactor/vutils';
 import { traverseGroup } from '../../util';
