+++ packages/vrender-core/src/interface/animation/index.ts
@@ -1,0 +1,5 @@
+export * from './animate';
+export * from './ticker';
+export * from './timeline';
+export * from './type';
+export * from './easing';
