+++ packages/vrender-core/src/interface/animation/type.ts
