+++ packages/vrender-core/src/interface/index.ts
@@ -73,4 +73,4 @@ export * from './plugin';
 export * from './picker';
 export * from './text';
 export * from './window';
-export * from './animate';
+export * from './animation';
